Rename `objects` lib as `implementations`
We called these things objects before because they were The Objects(TM).
However, now these objects are just default/example implementations of
the protocols. We've even begun referring to them as "implementations".
As such, it seemed time to rename them accordingly.
