+++ dbt_semantic_interfaces/implementations/elements/entity.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 from typing import Optional
 
-from dbt_semantic_interfaces.objects.base import (
+from dbt_semantic_interfaces.implementations.base import (
     HashableBaseModel,
     ModelWithMetadataParsing,
 )
-from dbt_semantic_interfaces.objects.metadata import Metadata
+from dbt_semantic_interfaces.implementations.metadata import Metadata
 from dbt_semantic_interfaces.references import EntityReference
 from dbt_semantic_interfaces.type_enums.entity_type import EntityType
