+++ dbt_semantic_interfaces/implementations/semantic_manifest.py
@@ -1,8 +1,8 @@
 from typing import List
 
-from dbt_semantic_interfaces.objects.base import HashableBaseModel
-from dbt_semantic_interfaces.objects.metric import Metric
-from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
+from dbt_semantic_interfaces.implementations.base import HashableBaseModel
+from dbt_semantic_interfaces.implementations.metric import Metric
+from dbt_semantic_interfaces.implementations.semantic_model import SemanticModel
 
 
 class SemanticManifest(HashableBaseModel):
