+++ dbt_semantic_interfaces/model_transformer.py
@@ -2,7 +2,7 @@ import copy
 import logging
 from typing import Sequence, Tuple
 
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.transformations.add_input_metric_measures import (
     AddInputMetricMeasuresRule,
 )
