+++ dbt_semantic_interfaces/protocols/semantic_model.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from abc import abstractmethod
 from typing import List, Optional, Protocol, Sequence
 
-from dbt_semantic_interfaces.objects.elements.dimension import Dimension
+from dbt_semantic_interfaces.implementations.elements.dimension import Dimension
 from dbt_semantic_interfaces.protocols.entity import Entity
 from dbt_semantic_interfaces.protocols.measure import Measure
 from dbt_semantic_interfaces.references import (
