+++ dbt_semantic_interfaces/test_utils.py
@@ -5,14 +5,21 @@ from typing import Callable, Optional, Sequence, Tuple
 
 import dateutil.parser
 
-from dbt_semantic_interfaces.objects.elements.dimension import Dimension
-from dbt_semantic_interfaces.objects.elements.entity import Entity
-from dbt_semantic_interfaces.objects.elements.measure import Measure
-from dbt_semantic_interfaces.objects.filters.where_filter import WhereFilter
-from dbt_semantic_interfaces.objects.metadata import FileSlice, Metadata
-from dbt_semantic_interfaces.objects.metric import Metric, MetricType, MetricTypeParams
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
-from dbt_semantic_interfaces.objects.semantic_model import NodeRelation, SemanticModel
+from dbt_semantic_interfaces.implementations.elements.dimension import Dimension
+from dbt_semantic_interfaces.implementations.elements.entity import Entity
+from dbt_semantic_interfaces.implementations.elements.measure import Measure
+from dbt_semantic_interfaces.implementations.filters.where_filter import WhereFilter
+from dbt_semantic_interfaces.implementations.metadata import FileSlice, Metadata
+from dbt_semantic_interfaces.implementations.metric import (
+    Metric,
+    MetricType,
+    MetricTypeParams,
+)
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.semantic_model import (
+    NodeRelation,
+    SemanticModel,
+)
 from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
 
 logger = logging.getLogger(__name__)
