+++ dbt_semantic_interfaces/transformations/convert_median.py
@@ -1,8 +1,8 @@
 from dbt_semantic_interfaces.errors import ModelTransformError
-from dbt_semantic_interfaces.objects.elements.measure import (
+from dbt_semantic_interfaces.implementations.elements.measure import (
     MeasureAggregationParameters,
 )
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.transformations.transform_rule import ModelTransformRule
 from dbt_semantic_interfaces.type_enums.aggregation_type import AggregationType
