+++ dbt_semantic_interfaces/validations/element_const.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 from typing import DefaultDict, List
 
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import SemanticModelReference
 from dbt_semantic_interfaces.validations.validator_helpers import (
     FileContext,
