+++ dbt_semantic_interfaces/validations/validator_helpers.py
@@ -12,9 +12,9 @@ import click
 from pydantic import BaseModel, Extra
 
 from dbt_semantic_interfaces.enum_extension import assert_values_exhausted
-from dbt_semantic_interfaces.objects.base import FrozenBaseModel
-from dbt_semantic_interfaces.objects.metadata import Metadata
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.base import FrozenBaseModel
+from dbt_semantic_interfaces.implementations.metadata import Metadata
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import (
     MetricModelReference,
     SemanticModelElementReference,
