+++ tests/validations/test_configurable_rules.py
@@ -2,13 +2,13 @@ import copy
 
 import pytest
 
-from dbt_semantic_interfaces.model_validator import ModelValidator
-from dbt_semantic_interfaces.objects.metric import (
+from dbt_semantic_interfaces.implementations.metric import (
     MetricInput,
     MetricType,
     MetricTypeParams,
 )
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.model_validator import ModelValidator
 from dbt_semantic_interfaces.test_utils import metric_with_guaranteed_meta
 from dbt_semantic_interfaces.validations.metrics import DerivedMetricRule
