+++ tests/validations/test_metrics.py
@@ -1,18 +1,18 @@
 import pytest
 
-from dbt_semantic_interfaces.model_validator import ModelValidator
-from dbt_semantic_interfaces.objects.elements.dimension import (
+from dbt_semantic_interfaces.implementations.elements.dimension import (
     Dimension,
     DimensionTypeParams,
 )
-from dbt_semantic_interfaces.objects.elements.entity import Entity
-from dbt_semantic_interfaces.objects.elements.measure import Measure
-from dbt_semantic_interfaces.objects.metric import (
+from dbt_semantic_interfaces.implementations.elements.entity import Entity
+from dbt_semantic_interfaces.implementations.elements.measure import Measure
+from dbt_semantic_interfaces.implementations.metric import (
     MetricInput,
     MetricType,
     MetricTypeParams,
 )
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
+from dbt_semantic_interfaces.model_validator import ModelValidator
 from dbt_semantic_interfaces.references import (
     DimensionReference,
     EntityReference,
