+++ tests/validations/test_unique_valid_name.py
@@ -3,8 +3,8 @@ from copy import deepcopy
 import more_itertools
 import pytest
 
+from dbt_semantic_interfaces.implementations.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.model_validator import ModelValidator
-from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.test_utils import find_semantic_model_with
 from dbt_semantic_interfaces.validations.unique_valid_name import (
     MetricFlowReservedKeywords,
